dev_requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev_requirements.in
#
beautifulsoup4==4.6.0     # via wagtail
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
django-bootstrap-form==3.4
django-extensions==2.2.1
django-modelcluster==3.1  # via wagtail
django-modeltranslation==0.13.3  # via wagtail-modeltranslation
django-rosetta==0.9.3
django-taggit==0.22.2     # via wagtail
django-treebeard==4.3     # via wagtail
django==1.11.23
djangorestframework==3.6.4  # via wagtail
html5lib==0.999999999     # via wagtail
idna==2.8                 # via requests
pillow==6.1.0             # via wagtail
polib==1.1.0              # via django-rosetta
psycopg2-binary==2.8.3
python-dotenv==0.10.3
pytz==2019.2              # via django, django-modelcluster
requests==2.22.0          # via django-rosetta, wagtail
six==1.12.0               # via django-extensions, django-rosetta, html5lib
unidecode==1.1.1          # via wagtail
urllib3==1.25.3           # via requests
wagtail-modeltranslation==0.10.2
wagtail==1.13.4
webencodings==0.5.1       # via html5lib
willow==1.0               # via wagtail

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via html5lib
